requirements.txt
pandas
numpy
scikit-learn
boto3
python-dotenv
flask
streamlit
joblib
sentence-transformers
lightgbm
PyPDF2
requests
accelerate>=0.26.0
-e .
